// UI Testing/cypress/integration/models/flipkart.ts
export class FlipkartSelectors {
    loginSelector: string = '._1_3w1N';
    mobileSelector: string = ':nth-child(1) > ._2IX_2-';
    passwordSelector: string = ':nth-child(2) > ._2IX_2-';
    loginButtonSelector: string = '._1D1L_j > ._2KpZ6l';
    loginCheckDataSelector: string = ':nth-child(3) > ._28p97w > ._1psGvi > :nth-child(1) > .exehdJ';
    searchSelector: string = '._3704LK';
    minRangeSelector: string = '._1YAKP4 > ._2YxCDZ';
    maxRangeSelector: string = '._3uDYxP > ._2YxCDZ';
    wishlistCartSelector: string = ':nth-child(3) > ._28p97w';
    totalItemSelector: string = '._3FVYY1 > span';
    loadItemSelector: string = '._2yhHWj';
    addToWishlistSelector: string = '._36FSn5';
    eachItemPriceSelector: string = '._30jeq3._1qGLe2';
    wishlistItemsSelector: string = '._1M-Ete';
    errorMessageSelector: string = '._2YULOR > span';
}